+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 3.5.2 (03/13/2024)
 * Fixed an issue where slide images are showing as black screen on Safari.
 * Allow YouTube iFrame embed to autoplay in Google Chrome when set as the HTML page type source.
+* Removed Vimeo as a page type. HTML page type can now accommodate Vimeo video embed.
 
 #### 3.5.1 (03/06/2024)
 * Downloadable file name is no longer reliant on the course directory or the URI path name. Downloadable file name can be set in the `sbplus.xml` file in the `storybook` node. If not specified, it defaults to the presentation title with spaces and special characters removed.